chore: update mocked data, change file extension to acept JSX elements
